# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ROUTB API",
    "description": "ROUTB",
    "version": "0.3.0"
  },
  "paths": {
...
    "/trips/{trip_id}": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get Trip",
        "operationId": "get_trip_trips__trip_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TripRequestSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "trip_id": {
            "type": "integer",
            "title": "Trip Id"
          },
          "passenger_id": {
            "type": "integer",
            "title": "Passenger Id"
          },
          "seat_count": {
            "type": "integer",
            "title": "Seat Count",
            "default": 1
          },
          "passenger_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passenger Name"
          },
          "passenger_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passenger Phone"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "trip_id",
          "passenger_id",
          "status",
          "created_at"
        ],
        "title": "TripRequestSummary"
      },
      "TripResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          },
          "total_seats": {
            "type": "integer",
            "title": "Total Seats"
          },
          "available_seats": {
            "type": "integer",
            "title": "Available Seats"
          },
          "departure_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departure Time",
            "default": "7:00 AM"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          },
          "driver_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Id"
          },
          "driver_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Name"
          },
          "my_request_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "My Request Status"
          },
          "requests": {
            "items": {
              "$ref": "#/components/schemas/TripRequestSummary"
            },
            "type": "array",
            "title": "Requests"
          }
        },
        "type": "object",
        "required": [
          "id",
          "origin",
          "destination",
          "total_seats",
          "available_seats"
        ],
        "title": "TripResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}